Added TimingPoint Support
I'm pretty happy on how the structure of TimingPoint turned out.

It's lightweight, simplistic, and straight to the point.

One thing I felt that I could improve is the usage of magic numbers in SliderVelocity and BPM Classes. There was a discrepency between static and nonstatic functions, so I had to split them.

I felt that it was too large of a stretch to workaround this, so I left it as it is, it's not very susceptible to errors anyways.